+++ src/main/java/bossmonster/Main.java
@@ -3,12 +3,13 @@ package bossmonster;
 import java.util.Scanner;
 
 import bossmonster.controller.Controller;
+import bossmonster.domain.monster.BossMonsterDamageGenerator;
 
 public class Main {
     public static void main(String[] args) {
         Controller controller = new Controller(
                 new Scanner(System.in),
-                new RandomBossMonsterDamageGenerator(),
+                new BossMonsterDamageGenerator(),
                 new Service()
         );
         controller.start();
