+++ src/main/java/bossmonster/Service.java
@@ -2,8 +2,13 @@ package bossmonster;
 
 import java.util.List;
 
-import bossmonster.monster.BossMonster;
-import bossmonster.player.Player;
+import bossmonster.domain.info.Hp;
+import bossmonster.domain.info.Mp;
+import bossmonster.domain.info.Name;
+import bossmonster.domain.info.Stat;
+import bossmonster.domain.monster.BossMonster;
+import bossmonster.domain.player.Attack;
+import bossmonster.domain.player.Player;
 
 public class Service {
