+++ src/main/java/bossmonster/domain/info/Hp.java
@@ -1,4 +1,4 @@
-package bossmonster;
+package bossmonster.domain.info;
 
 public class Hp {
