+++ src/main/java/bossmonster/domain/info/Mp.java
@@ -1,4 +1,4 @@
-package bossmonster;
+package bossmonster.domain.info;
 
 import bossmonster.utils.ErrorMessage;
